backend/requirements.txt
bandit>=1.9.4
fastapi>=0.135.1
httpx[http2]>=0.28.1
pydantic>=2.12.5
pytest>=9.0.2
pytest-asyncio>=1.3.0
python-dotenv>=1.2.2
python-multipart>=0.0.22
ruff>=0.15.6
slowapi>=0.1.9
uvicorn>=0.41.0
